requirements.txt
# Web Framework
flask==2.3.3
flask-cors==4.0.0

# AI/ML Libraries
scikit-learn==1.3.0
nltk==3.8.1
spacy==3.7.2
numpy==1.24.3
pandas==1.5.3

# Database
sqlalchemy==2.0.21

# Data Processing
requests==2.31.0
python-dateutil==2.8.2

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Development Tools
black==23.7.0
flake8==6.0.0
python-dotenv==1.0.0

# Environment Variables
python-decouple==3.8

# API Documentation
flasgger==0.9.7.1

# Logging
structlog==23.1.0

# Data Validation
marshmallow==3.20.1